requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.2.1
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.5
fastapi-socketio==0.0.10
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.10.1
pydantic_core==2.27.1
python-engineio==4.10.1
python-jose==3.3.0
python-multipart==0.0.17
python-socketio==5.11.4
requests==2.32.3
rsa==4.9
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wsproto==1.2.0
